requirements.txt
Django==1.3.1
boto==2.2.2
cssmin==0.1.4
django-appconf==0.5
django-pipeline==1.2b6
django-staticfiles==1.2.1
django-storages==1.1.4
